==> wrangle/00_business_cleanup.ipynb <==
# Yelp Businesses: Cleaning and Wrangling

The objective of this notebook is to inspect and wrangle the `business.json` file from the yelp dataset.
At each feature extracting/cleaning step, the data is saved in a separate csv file in the format `business_feature.csv` such that we can trace back the file origin. This is also to avoid ending up with a massing dataframe with too many features.


import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import os
import seaborn as sns
from collections import Counter, OrderedDict
import calendar
import utils

%matplotlib inline

# Load + Assess

#location of file
business_dir = 'data/business.json'

#download data
df_bus = utils.chunk_loader(business_dir, read_limit=-1)
#df_bus = pd.read_json(business_dir, orient='columns',lines=True)

#head
df_bus.head()

#shape
df_bus.shape

#data types
df_bus.dtypes

#quick stats
df_bus.describe()

#percent missing
df_bus.isna().mean()

# Cleaning to-do list

Evaluating the business dataframe above, there are several features that need to be cleaned.
The list below offers a roadmap to addressing these issues although it might not be comprehensive. During the process we might need to add additional steps.
We understand that some data types are nested within the columns, and that these data types might not be stored  in the appropirate manner.


- address
    - make everything lower case
    - extract feature: if on road/boulevard/ave/etc...
- attributes
    - break up dict to dummy variables
- business_id
    - no changes
- categories
    - make everything lower?
    - dummy variables and split by comma character
    - note that not everything is a restaurant (plumbers)
- city
    - maybe lower case?
- hours
    - split dict by days
        - open hour monday
        - close hour monday
        - etc...
    - figure out placeholder value for None
    - check if correlation between closed restaurant and no hours posted
- is_open
    - no changes
    - 82% are open, 18% are dead businesses
- latitude
    - no changes
- longitude
    - no changes
- name
    - no changes
- postal_code
    - replace zip code with integer representation
- review_count
    - note that lowest value is 3
- stars
    - no changes
- state
    - some are canadian
    - add feature: is in USA yes/no

#mapping dict for replacing and fixing data types
bool_to_int = {True: 1, False: 0, np.nan: 0, 'True': 1, 'False': 0, 'None': 0, None: 0}

## address

#create deep copy
df_bus_adr = df_bus.copy()
#make everything lower case
df_bus_adr['address'] = df_bus_adr['address'].str.lower()
#remove punctuation
df_bus_adr['address'] = df_bus_adr['address'].str.replace('[^\w\s]','')

#define counter object
adr_counter = Counter()
#loop over every address entry
for add in df_bus_adr.address:
    #loop over each individial word
    for word in add.lower().split():
        #add word to counter
        adr_counter[word] +=1

#list top k words
adr_counter.most_common(100)

#road type mapping to homogenize road names
road_type_dict = {'rd': 'road','rue': 'road', 'avenue': 'ave',
                  'street': 'str', 'blvd': 'boulevard',
                  'drive': 'dr', 'highway': 'hwy',
                  'parkway': 'pkwy', 'center': 'ct', 'lane': 'ln'}

#replace names
df_bus_adr['address'] = df_bus_adr['address'].replace(road_type_dict, regex=True)

#get list of finalized road values
#set to remove duplicates
road_types_list = list(set(road_type_dict.values()))
print(road_types_list)

#dict for dummies
road_col_dict = {}
#iterate over road types
for road in road_types_list:
    #create a dummy for that type
    dum_col = df_bus_adr['address'].str.contains(road)
    #add it to the dict
    road_col_dict[road] = dum_col

#convert boolean to 1/0
road_type_df = pd.DataFrame.from_dict(road_col_dict).replace({False:0, True:1})

road_type_df.head()

#key for reference in case we want to join tables
road_type_df['business_id'] = df_bus['business_id']

#save the work
road_type_df.to_csv(path_or_buf='data/cleaned/business_roadtype.csv')

### Alternatively we can create a categorical series instead of one hot encoding to reduce the dimensionality of the combined dataframe. We can do this because the categories are mutually exclusive.

#drop index col since we use dataframe later down
road_type_df = road_type_df.drop(columns=['business_id'])

#create dict mapping for road types
road_to_cat_dict = dict(zip(road_type_df.columns, np.arange(1, len(road_type_df.columns)+1)))

#stak dataframe
road_type_df_stack = road_type_df.stack()

#convert to series
road_type_df_cats = pd.Series(pd.Categorical(road_type_df_stack[road_type_df_stack!=0].index.get_level_values(1)))

#replace with dict
road_type_df_cats = road_type_df_cats.replace(road_to_cat_dict)


road_type_df_cats.head()

road_type_df_cats.shape

df_bus['road_type'] = road_type_df_cats

#fill missing values with 0
df_bus['road_type'] = df_bus['road_type'].fillna(0)

### We can't necessarily rank the importance of the address based on its designation. A road might be more visited than a parkway, and the numbering of the road might not indicate anything. Thinking ahead, we believe that using one hot encoding might be the best solution for this categorical variable

### Replace postal codes with corresponding integer code

### Alternatively:
use IRS data to get income by zip code

https://www.irs.gov/statistics/soi-tax-stats-individual-income-tax-statistics-2016-zip-code-data-soi

https://www.irs.gov/statistics/soi-tax-stats-individual-income-tax-statistics-zip-code-data-soi


For Canada:
https://www12.statcan.gc.ca/census-recensement/2016/dp-pd/hlt-fst/inc-rev/Table.cfm?Lang=Eng&T=102&PR=0&D1=1&RPP=25&SR=1&S=108&O=D



#get unique postal codes
unique_postal = df_bus['postal_code'].unique().tolist()

#map code to integer representation
postal_to_int = dict(zip(unique_postal, np.arange(0, len(unique_postal))))

#reverse mapping for reference
int_to_postal = {v:k for k,v in postal_to_int.items()}

#replace in dataframe
df_bus['postal_code'] = df_bus['postal_code'].replace(postal_to_int)

## attributes

#break up dict inside df
df_atr = df_bus['attributes'].apply(pd.Series)
df_atr.head(10)

#inspect types
df_atr.dtypes

#select cols that are still as objects
df_atr_obj_cols = df_atr.select_dtypes(include='object').columns
print(df_atr_obj_cols)

df_atr[df_atr_obj_cols].head(10)

def str_dict_to_df(series):
    """
    Takes in a pandas series with dicts stored as strings
    returns dataframe with dict keys as columns
    
    serires: pandas series
    """
    eval_list = []
    for sr in series:
        if not pd.isna(sr):
            eval_list.append(eval(sr))
        else:
            eval_list.append(np.nan)
    
    eval_df = pd.Series(eval_list).apply(pd.Series)  
    
    #drop cols that are all nan
    eval_df = eval_df.dropna(axis=1, how='all')
    
    
    return eval_df

#store col names
dict_cols_list = []

for col in df_atr_obj_cols:
    #if contains a curly bracket, then assume column is a dict as string
    if df_atr[col].str.contains('{').any():
        dict_cols_list.append(col)
print(dict_cols_list)            

#store dataframes from dict nested columns
dict_col_df_list = []

for col in dict_cols_list:
    #apply string to dict evaluation
    temp_df = str_dict_to_df(df_atr[col])
    #append to list
    dict_col_df_list.append(temp_df)

#combine all in one column
dict_col_df = pd.concat(dict_col_df_list, axis=1)
dict_col_df.head()

#store dummy attribute columns in list
atr_dum_df_cols = []

#iterate over object columns
for col in df_atr_obj_cols:
    #if contains a categorical variable then it begins with a u
    if df_atr[col].str.contains("u'").any():
        #add to the list
        atr_dum_df_cols.append(col)

#print out list
print(atr_dum_df_cols)

#explore unique entries in attribute columns
for col in atr_dum_df_cols:
    print(col)
    print(df_atr[col].unique())
    print(50*'-')


### We build a dictionary to replace each categorical variable with a numerical value that corresponds to a sliding scale. For example we can rank attire from casual, to dressy, and to formal. This will reduce the feature space, and probably provide more information gain than a one hot encoding. For simplicity we will assume that a NaN is always the lowest value on the scale.

#noise level
noise_level_dict = dict(zip([np.nan, "u'loud'" ,"u'average'" ,
                             "'average'" ,"u'quiet'" ,"'loud'" ,
                             "'very_loud'", "'quiet'" ,"u'very_loud'" ,'None'], 
                            [0, 3, 2, 
                             2, 1, 3, 
                             4, 1 , 4, 0]))
print(noise_level_dict)




#free wifi is the best kind of wifi
wifi_dict = dict(zip(
    [np.nan ,"u'no'" ,"'free'" ,
     "u'free'" ,"'no'" ,"u'paid'" ,"'paid'" ,'None'], 
    [0, 1, 3, 
     3, 1, 2, 2, 0]))
print(wifi_dict)

#more alcohol is better
alcohol_dict = dict(zip([np.nan ,"u'full_bar'" ,"u'beer_and_wine'" ,
                         "u'none'" ,"'beer_and_wine'" ,"'none'", 
                         "'full_bar'" ,'None'],
                        [0, 2, 1, 
                         0, 1, 0, 
                         2, 0]))
print(alcohol_dict)

#rank by formality
restaurant_attire_dict = dict(zip(
    [np.nan ,"u'casual'" ,"'casual'" ,
     "'dressy'" ,"u'dressy'" ,"u'formal'" ,
     'None',"'formal'"], 
    [0, 1, 1, 
     2, 2, 3, 
     0, 3]))
print(restaurant_attire_dict)

#free corkage is best
byob_corkage_dict = dict(zip(
    [np.nan ,"'no'" ,"'yes_corkage'" ,
     "'yes_free'" ,"u'no'" ,"u'yes_free'","u'yes_corkage'"],
    [0, 1, 2, 
     3, 1, 3, 2]))
print(byob_corkage_dict)

#from least to most accomodating for smokers
smoking_dict = dict(zip(
    [np.nan ,"u'no'" ,"u'outdoor'" ,
     "u'yes'" ,"'no'" ,'None' ,
     "'outdoor'" ,"'yes'"], 
    [0, 1, 2, 
     3, 1, 0, 
     2, 3]))
print(smoking_dict)

#from youngest to oldest crowd
ages_allowed_dict = dict(zip(
    [np.nan ,"u'allages'" ,"u'21plus'" ,
     "u'18plus'" ,"u'19plus'" ,'None'], 
    [0, 1, 4, 
     2, 3, 0]))
print(ages_allowed_dict)

#define mapping for replace method
atr_dict_map = {'NoiseLevel': noise_level_dict, 
                'WiFi': wifi_dict, 
                'Alcohol': alcohol_dict, 
                'RestaurantsAttire': restaurant_attire_dict, 
                'BYOBCorkage': byob_corkage_dict, 
                'Smoking': smoking_dict, 
                'AgesAllowed': ages_allowed_dict}

#get a copy with relevant columns
df_atr_categorical = df_atr[atr_dum_df_cols].copy()
#apply replacement
df_atr_categorical = df_atr_categorical.replace(atr_dict_map)

df_atr_categorical.head()

#### The cell block below will return a one hot encoding of the categorical attributes above, uncomment if objective changes.

# #store dummy dataframes in list
# atr_dum_df_list = []

# for col in atr_dum_df_cols:
    
#     #we do not want to modify the dataframe in place so create a copy
#     temp_series = df_atr[col].copy()
    
    
#     #fx messy inputs and remove u
#     temp_series = temp_series.str.replace("u'", "")
#     #remove '
#     temp_series= temp_series.str.replace("'", "")
        
    
#     #create dummies
#     dum_df = pd.get_dummies(temp_series)
#     #drop the None column
#     dum_df = dum_df.drop(columns=['None'])
        
#     #fx messy col names in case we missed them
#     dum_df.columns = dum_df.columns.str.replace("u'", "")
#     dum_df.columns = dum_df.columns.str.replace("'", "")
        
#     #add prefixt
#     dum_df = dum_df.add_prefix(col+'_')
                
#     #append to list
#     atr_dum_df_list.append(dum_df)

# #concat
# atr_dum_df = pd.concat(atr_dum_df_list, axis=1)

# atr_dum_df.head()


# #standard preprocessing for restaurant price range since it has no u
# atr_price_range = pd.get_dummies(df_atr['RestaurantsPriceRange2'])
# atr_price_range = atr_price_range.drop(columns= ['None'])
# atr_price_range = atr_price_range.add_prefix('price_range'+'_')

# atr_price_range.head()

#add to the list
atr_to_drop =  dict_cols_list + list(df_atr_categorical.columns)
print(atr_to_drop)

#combine features in one dataframes
df_atr_conc = pd.concat([df_atr.drop(columns=atr_to_drop),dict_col_df, df_atr_categorical], axis=1)

#standardize name
df_atr_conc.columns = df_atr_conc.columns.str.replace("-", "_")

#make 1/0
df_atr_conc = df_atr_conc.replace(bool_to_int)

df_atr_conc.head()

#quick check for column names
df_atr_conc.columns

#key for reference
df_atr_conc['business_id'] = df_bus['business_id']

#save the work
df_atr_conc.to_csv(path_or_buf='data/cleaned/business_attributes.csv')

## Categories

#counter for original categories
cat_counter = Counter()
#loop through split categories
for cat_split in df_bus['categories'].str.split(',| '):
    #if statement to avoid none type is not iterable
    if cat_split:
        for cat in cat_split:
            cat_counter[cat] +=1

#see top k types
cat_counter.most_common(25)

#get keys for top k common categories
top_cats = list(dict(cat_counter.most_common(25)).keys())
print(top_cats)
#note that the first 2 were space and & so skip those
print("\nselecting only top 10 relevant sections\n")
top_cats = top_cats[3:13]
print(top_cats)

#create dict for categories
category_dict = {}
for cat in top_cats:
    #select relevant top 10 categories from before
    dum_cat = df_bus['categories'].str.contains(cat)
    #add it to the dict
    category_dict[cat] = dum_cat
    
cat_type_df = pd.DataFrame.from_dict(category_dict).replace(bool_to_int)

cat_type_df.head()

#key for reference
cat_type_df['business_id'] = df_bus['business_id']

#save the work
cat_type_df.to_csv(path_or_buf='data/cleaned/business_cats.csv')

## Hours

Holy grail of date time format:

http://strftime.org/

#split out the dict
hours_day_df = df_bus['hours'].apply(pd.Series)
hours_day_df.head()

def series_to_datetime(df, series):
    """
    Takes in a pandas series with content stored as a string
    must have format hour:minute - hour:minute
    can have missing values
    
    df = pandas dataframe
    series = pandas column name
    """
    #create array for day of week
    #weekday = list(calendar.day_abbr)
    
    #ordered dict container
    serires_dict = OrderedDict()
    
    #hour container
    open_hour = []
    close_hour = []
    
    #split the series along the dash (-)
    day = df[series].str.split("-")

    #iterate over days
    for hour in day:
        
        #if not a nan the split will return a list
        if type(hour)==list:
            open_hour.append(hour[0])
            close_hour.append(hour[1])
        else:
            #necessary nan for when not available
            open_hour.append(np.nan)
            close_hour.append(np.nan)
            
    #make a datetime object    
    
    open_hour_dt = pd.to_datetime(open_hour, dayfirst=True,format='%H:%M')
    close_hour_dt = pd.to_datetime(close_hour, dayfirst=True, format='%H:%M')
    
    serires_dict[series+'_open'] = open_hour
    serires_dict[series+'_close'] = close_hour
    
    hours_df = pd.DataFrame.from_dict(serires_dict)
    
    return hours_df

hours_df_list = []

for col in hours_day_df.columns:
    temp_hour_df = series_to_datetime(hours_day_df, col)
    hours_df_list.append(temp_hour_df)
    
    
hours_df_openclose = pd.concat(hours_df_list, axis=1)
hours_df_openclose.head()

hours_df_openclose.dtypes

#key for reference
hours_df_openclose['business_id'] = df_bus['business_id']

#save the work
hours_df_openclose.to_csv(path_or_buf='data/cleaned/business_hours.csv')

## is_open

#quick check for unique values
df_bus.is_open.unique()

## state

#USA! USA!

states = ["AL", "AK", "AZ", "AR", "CA", "CO", "CT", "DC", "DE", "FL", "GA", 
          "HI", "ID", "IL", "IN", "IA", "KS", "KY", "LA", "ME", "MD", 
          "MA", "MI", "MN", "MS", "MO", "MT", "NE", "NV", "NH", "NJ", 
          "NM", "NY", "NC", "ND", "OH", "OK", "OR", "PA", "RI", "SC", 
          "SD", "TN", "TX", "UT", "VT", "VA", "WA", "WV", "WI", "WY"]

#get usa sates and make dataframe
df_is_usa = pd.DataFrame(df_bus['state'].str.contains('|'.join(states)).astype(int))
df_is_usa['business_id'] = df_bus['business_id']
#drop redundant column
df_is_usa = df_is_usa.drop(columns=['state'])

#save the work
df_is_usa.to_csv(path_or_buf='data/cleaned/is_usa.csv', header=True)

### Combine DFs in master DF with business_id as common key

# #superseded in favor for utils.reduce_merge

# #combine in massive df
# df_list = [df_bus, 
#            road_type_df_cats, 
#            df_atr_conc.drop(columns=['business_id']), 
#            cat_type_df.drop(columns=['business_id']), 
#            hours_df_openclose.drop(columns=['business_id']), 
#            df_is_usa.drop(columns=['business_id'])]

# df_bus_conc = pd.concat(df_list, axis=1)

# df_bus_conc.shape

#combine in massive df
df_list = [df_bus, 
           df_atr_conc, 
           cat_type_df, 
           hours_df_openclose, 
           df_is_usa]

df_bus_conc = utils.reduce_merge(df_list, key='business_id')

df_bus_conc.shape

#check columns
print(*df_bus_conc.columns, sep=',   ')

#save the work
df_bus_conc.to_csv(path_or_buf='data/cleaned/df_bus_conc.csv', header=True)
