==> wrangle/02_CensuData.ipynb <==
# Use External Databases

##### In this notebook we make use of external sources such as US IRS data and the Canadian Census Bureau to turn zip codes into meaningful data, in particular the **household median income**. 

https://www.irs.gov/statistics/soi-tax-stats-individual-income-tax-statistics-2016-zip-code-data-soi

https://www.irs.gov/statistics/soi-tax-stats-individual-income-tax-statistics-zip-code-data-soi

For Canada: 

https://www12.statcan.gc.ca/census-recensement/2016/dp-pd/hlt-fst/inc-rev/Table.cfm?Lang=Eng&T=102&PR=0&D1=1&RPP=25&SR=1&S=108&O=D


Understanding Canadian postal codes:

https://www.postalcodesincanada.com/

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import utils

%matplotlib inline

# United States of America

irs = utils.chunk_loader('census_data/16zpallagi.csv', index_col=False)

irs.head()

irs.shape

#checking my address for fun
irs[irs['zipcode']==94104]

irs.columns

From the dataframe documentation we care about two features:
- zipcode
- AGI stub: Size of adjusted gross income
    - 1 = 1 under 25,000
    - 2 = 25,000 under 50,000
    - 3 = 50,000 under 75,000
    - 4 = 75,000 under 100,000
    - 5 = 100,000 under 200,000
    - 6 = 200,000 or more
Note that all incomes reported here are in United States dollar.

We can use this to build a sliding scale and assign a median household income for each zip code. The rationale is that businesses definetly need to know what the income is in their area of operation in order to study their business model, products, pricing, etc...

irs_median = irs.groupby(by=['zipcode'], as_index=True).median()
irs_median = irs_median['agi_stub']
irs_median.head()

irs_median.unique()

#conver to dict
usa_median = irs_median.to_dict()

#get mean:
usa_median['mean'] = irs_median.mean()

# Canada

For Canada some differences with USA:
- Data reported in Canadian dollars
- Use 0.76 Canadian dollars for each USA dollar
- Postal code format differs
- Data might be less granular
- Encoding might differ
- Consistent with IRS take income before taxes

The Canadian data is a lot less granular than USA however even at the provincial level we know that income will matter a lot.

#get encoding of file via commandline
! chardetect census_data/canada/98-402-X2016006-T1-CSD-Eng.csv

#download data
canada = pd.read_csv('census_data/canada/98-402-X2016006-T1-CSD-Eng.csv',
                     encoding='windows-1252')

#drop columns that have missing values
canada = canada.dropna(axis=1, how='any')

#rename relevant columns
canada = canada.rename(columns={'Geographic name, Province or territory': 'province', 
                                'Median household total income (2015 constant dollars), 2015': 'median_2015',
                                'Number of households, 2016': 'household_count_2016', 
                                'Number of households, 2005': 'household_count_2005'})


canada.head()

#check cities
canada['Geographic name'].unique()

#check provinces
canada['province'].unique()

canada.shape

canada.columns

canada_sigma = canada.groupby(by=['province'], as_index=True).sum()
canada_pop = canada_sigma['household_count_2016'].to_dict()
canada_pop

def weighted_income(df, 
                    prov, 
                    pop_dict=canada_pop, 
                    prov_col='province',
                    median_col='median_2015', 
                    pop_col='household_count_2016', 
                    conversion_rate=0.76):
    #get dataframe of province
    df_prov = df[df[prov_col]==prov].copy()
    
    #number of households in province
    population = pop_dict[prov]
    
    #get fraction contribution of each region
    df_prov['weights'] = df_prov[pop_col]/population
    
    #get weighted median
    weighted_median = (df_prov[median_col] * df_prov['weights']).sum()
    
    #convert to american dollars
    weighted_median *= conversion_rate
    
    return int(weighted_median)

def income_to_cat(income):
    if income <= 25000:
        return 1
    elif income <=50_000:
        return 2
    elif income <=75_000:
        return 3
    elif income <=100_000:
        return 4
    elif income <=200_000:
        return 5
    elif income >200_000:
        return 6

#test it out
print(weighted_income(canada, 'Yukon'))

canada_median = {}
canada_median_list = []

for province in canada_pop:
    median = income_to_cat(weighted_income(canada, province))
    canada_median[province] = median
    canada_median_list.append(median)
    
canada_median['mean'] = np.mean(canada_median_list)

canada_median

# Postal Codes to Median Incomes

def canada_post_to_prov(postal):
    first = postal[0]
    
    if first in 'A':
        return 'Newfoundland and Labrador'
    elif first in 'B':
        return 'Nova Scotia'
    elif first in 'C':
        return 'Prince Edward Island'
    elif first in 'E':
        return 'New Brunswick'
    elif first in 'GHJ':
        return 'Quebec'
    #note W is part of ontario
    elif first in 'KLMNPW':
        return 'Ontario'
    elif first in 'R':
        return 'Manitoba'
    elif first in 'S':
        return 'Saskatchewan'
    elif first in 'T':
        return 'Alberta'
    elif first in 'V':
        return 'British Columbia'
    elif first in 'X':
        return 'Northwest Territories'
    elif first in 'Y':
        return 'Yukon'
    else:
        return None

#location of file
business_dir = 'data/business.json'

#download businsess
df_bus = utils.chunk_loader(business_dir, read_limit=-1)
df_bus.head()

def postal_to_cat(postal, canada_median=canada_median, usa_median=usa_median):
    #check for missing values and if so return mean category for usa
    if len(postal) == 0:
        return usa_median['mean']
    
    elif not postal.isdigit():
        try:
            #make upper case in case of typo error
            province = canada_post_to_prov(postal[0].upper())
            return canada_median[province]
        #Some postal codes are UK
        except KeyError:
            return 0
    elif postal.isdigit():
        #note the type conversion
        try:
            median = usa_median[int(postal)]
            return median
        except KeyError:
            return usa_median['mean']
    else:
        return 0

#get reduced dataframe
df_bus_postal = df_bus[['business_id', 'postal_code']].copy()

#convert postal codes to categorical variable
df_bus_postal['median_income'] = df_bus_postal['postal_code'].apply(postal_to_cat)

#drop postal code after usage
df_bus_postal = df_bus_postal.drop(columns=['postal_code'])

df_bus_postal.head()

#save the work
df_bus_postal.to_csv('data/cleaned/business_median_income.csv')
