==> wrangle/31_checkin_feateng.ipynb <==
import pandas as pd
import numpy as np
import utils, plot_help
import matplotlib.pyplot as plt

#avoid warning popping up
from pandas.plotting import register_matplotlib_converters
register_matplotlib_converters()

%matplotlib inline

event_df = utils.chunk_loader('data/cleaned/checkin_eventlog.csv')
event_df.head()

### get first and last checkin of business and span over which checkins are being recorded.

#create a groupby object
event_df_groupby = event_df.reset_index(level=0).groupby(by='business_id', as_index=False)

#get first
first_checkin = event_df_groupby.min()
first_checkin = first_checkin.rename(columns={'index': 'first'})
first_checkin['first'] = pd.to_datetime(first_checkin['first']) 

first_checkin.head()

#get last
last_checkin = event_df_groupby.max()
last_checkin = last_checkin.rename(columns={'index': 'last'})
last_checkin['last'] = pd.to_datetime(last_checkin['last']) 

last_checkin.head()

first_last_df = pd.merge(left=first_checkin, 
                         right=last_checkin, 
                         how='inner',
                         left_on='business_id', 
                         right_on='business_id')

#get difference between first and last as seconds
first_last_df['span_checkin'] = (first_last_df['last'] - first_last_df['first']).apply(lambda x: x.seconds)

first_last_df.head()

### Evaluate what is the average monthly checkin for each business

event_df_month = event_df.copy()
event_df_month['month'] = pd.DatetimeIndex(event_df.index).month
event_df_month['ones'] = 1

event_df_month.head()

#get total checkin by month
checkin_month_count = event_df_month.groupby(by=['business_id', 'month'], as_index=False).sum()
checkin_month_count.head()

#group by business ID to get average monthly checkin
checkin_month_avg = checkin_month_count.groupby(by='business_id', as_index=False).mean()
checkin_month_avg.head()

#combine in single dataframe
df_checkin = pd.merge(left=checkin_month_avg, 
                      right = first_last_df, 
                      how='inner', 
                      on='business_id')

df_checkin = df_checkin.drop(columns=['month'])

df_checkin = df_checkin.rename(columns={'ones': 'avg_month_checkin'})

df_checkin.head()

df_checkin.to_csv('data/cleaned/checkin_feat.csv')
