==> wrangle/50_combine.ipynb <==
### The objective of this notebook is to combine the previous dataframes into one major dataframe.

import pandas as pd
import numpy as np
import utils, plot_help
import matplotlib.pyplot as plt

#avoid warning popping up
from pandas.plotting import register_matplotlib_converters
register_matplotlib_converters()

%matplotlib inline

list of dataframes to merge:
- df_bus_conc.csv
- review_year_change.csv
- checkin_feat.csv

# Businesses

df_bus_conc = utils.chunk_loader('data/cleaned/df_bus_conc.csv')
df_bus_conc.head()

df_bus_conc.shape

# Reviews

review_year_change = utils.chunk_loader('data/cleaned/review_year_change.csv')
review_year_change = review_year_change.rename(columns= {'index': 'business_id'})
review_year_change.head()

review_year_change.shape

# Checkins

checkin_feat = utils.chunk_loader('data/cleaned/checkin_feat.csv')
checkin_feat.head()

checkin_feat.shape

# Median Income

df_median = utils.chunk_loader('data/cleaned/business_median_income.csv')

df_median.head()

# combine dataframes in single df

df_list = [df_bus_conc, review_year_change, checkin_feat, df_median]

df_merge = utils.reduce_merge(df_list, 'business_id')
df_merge.head()

#get list of columns
cols = list(df_merge.columns.values)
#remove is_open from list
cols.pop(cols.index('is_open'))
#move is_open to last position
df_merge= df_merge[cols + ['is_open']]

print(*df_merge.columns, sep=',  ')

df_merge.to_csv('data/cleaned/business_merge_feats.csv')

def move_col_to_end(df, columns):
    """
    Returns a copy of dataframe with a specified column moved to the end
    """
